Testing settup

+++ src/main/java/com/example/taskmanager/service/TaskService.java
@@ -24,6 +24,8 @@ public class TaskService {
     }
 
     public void deleteTask(Long id) {
-        taskRepository.deleteById(id);
+        if (taskRepository.existsById(id)) {
+            taskRepository.deleteById(id);
+        }
     }
 }
